Atualização

+++ validacoes.h
@@ -196,7 +196,7 @@ string lerTextoDoTeclado(int quantidade) {
 	} while (retornoDaLeitura == 0);
 
 	//Validar tamanho da string lida
-	if (strlen(texto) > quantidade) {
+	if ((int)strlen(texto) > quantidade) {
 
 		lancarErro(17);
 		perror(MENSAGEM_DE_ERRO);
